# server-server/src/main/resources/static/openapi.json
{
  "openapi" : "3.0.0",
  "servers" : [
    {
      "url" : "https://api.draco.menkalian.de",
      "description" : "Default public production server"
    },
    {
      "url" : "http://localhost:8080",
      "description" : "Your local running instance"
    }
  ],
  "info" : {
    "description" : "`draco` is a project to develop an application to play the game \"quizpoker\".\nPart of that project is the creation of an database for so called \"guesstimate\"-questions.\nThis database is the part which is probably most useful to other applications. \n\nThis file shows the API-Endpoint which can be integrated in your application.\nThere is no API-Token required for the most usefull read-endpoints, but please try to keep the requests on a reasonable level.\n\nThe project also provides a `client-baseclient`-module which compiles to JVM and native code, so you can use it directly in your applications.",
    "version" : "1.0.0",
    "title" : "Draco",
    "termsOfService" : "https://draco.menkalian.de/legal/tos",
    "license" : {
      "name" : "MIT",
      "url" : "https://opensource.org/licenses/MIT"
    }
  },
...
  "paths" : {
...
    "/user" : {
      "put" : {
        "summary" : "Create a new user",
        "description" : "",
        "parameters" : [],
        "operationId" : "createUser",
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntitledUser"
                }
              }
            }
          },
          "401" : {
            "description" : "No Authentication found or user not authorized"
          }
        },
        "tags" : [
          "user"
        ],
        "security" : [
          {
            "auth_token" : []
          }
        ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EntitledUser"
              }
            }
          }
        }
      }
    },
    "/user/all" : {
      "get" : {
        "summary" : "Reads all users",
        "description" : "",
        "parameters" : [],
        "operationId" : "readAllUsers",
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/EntitledUser"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "No Authentication found or user not authorized"
          }
        },
        "tags" : [
          "user"
        ],
        "security" : [
          {
            "auth_token" : []
          }
        ]
      }
    },
    "/user/{id}" : {
      "post" : {
        "summary" : "Update an user",
        "description" : "",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            },
            "description" : "Id of the user"
          }
        ],
        "operationId" : "updateUser",
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntitledUser"
                }
              }
            }
          }
        },
        "tags" : [
          "user"
        ],
        "security" : [
          {
            "auth_token" : []
          }
        ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EntitledUser"
              }
            }
          }
        }
      },
      "get" : {
        "summary" : "Read data of an user",
        "description" : "",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            },
            "description" : "Id of the user"
          }
        ],
        "operationId" : "readUser",
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntitledUser"
                }
              }
            }
          }
        },
        "tags" : [
          "user"
        ],
        "security" : [
          {
            "auth_token" : []
          }
        ]
      },
      "delete" : {
        "summary" : "Deletes an user",
        "description" : "",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            },
            "description" : "Id of the user"
          }
        ],
        "operationId" : "deleteUser",
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "boolean"
                }
              }
            }
          },
          "401" : {
            "description" : "No Authentication found or user not authorized"
          }
        },
        "tags" : [
          "user"
        ],
        "security" : [
          {
            "auth_token" : []
          }
        ]
      }
    },
    "/user/me" : {
      "get" : {
        "summary" : "Read current user",
        "description" : "",
        "parameters" : [],
        "operationId" : "getMyUser",
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntitledUser"
                }
              }
            }
          },
          "401" : {
            "description" : "No valid user authenticated"
          }
        },
        "tags" : [
          "user"
        ],
        "security" : [
          {
            "auth_token" : []
          }
        ]
      }
    }
  },
...
  "components" : {
    "schemas" : {
...
      "UserRight" : {
        "type" : "string",
        "enum" : [
          "ADMIN",
          "USER_READ",
          "USER_CREATE",
          "USER_UPDATE",
          "USER_DELETE",
          "SUGGESTION_READ",
          "SUGGESTION_UPDATE",
          "SUGGESTION_DELETE",
          "SUGGESTION_COMMENT_CREATE",
          "QUESTION_READ",
          "QUESTION_CREATE",
          "QUESTION_UPDATE",
          "QUESTION_DELETE"
        ]
      },
      "EntitledUser" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "accessHash" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "rights" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserRight"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "auth_token" : {
        "type" : "apiKey",
        "name" : "auth",
        "in" : "header"
      }
    },
...
  },
  "security" : []
}